requirements.txt
# Core
fastapi
uvicorn[standard]
websockets
python-dotenv
pydantic

# Audio capture & VAD
sounddevice
numpy
scipy
# silero-vad loaded via torch.hub at runtime

# Transcription (Cohere Transcribe 2B)
# torch + torchaudio must be installed separately — RTX 5090 requires nightly cu130:
#   pip install --pre "torch>=2.7" "torchaudio>=2.7" --index-url https://download.pytorch.org/whl/nightly/cu130
transformers>=5.4.0
accelerate
librosa
sentencepiece
protobuf

# Speaker embedding (for enrolled speaker recognition)
pyannote.audio

# LM Studio / LLM
openai

# Database
aiosqlite
sqlalchemy

# Utilities
python-multipart
aiofiles
pystray
Pillow
rich
